ai-scaling-achitechture/ingestion-api/requirements.txt
Flask==2.2.2
google-cloud-pubsub==2.18.0
Werkzeug==2.2.2
google-cloud-storage==2.7.0
